+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [next] - unreleased
 
-- [#<PR-NUMBER>](https://github.com/simonneutert/papierkram_api_client/pull/<PR-NUMBER>) PR description. [@<gh-username>](https://github.com/<gh-username>)
+- [#90](https://github.com/simonneutert/papierkram_api_client/pull/90) crud ops for time_entries. [@simonneutert](https://github.com/simonneutert)
 
 ## [0.3.1] - 2023-10-05